refactor: clean ListManagerService
Still some things to do, but refactoring should be fine now.
