+++ src/app/model/garland-tools/item-data.ts
@@ -1,8 +1,5 @@
 import {Item} from './item';
 import {Partial} from './partial';
-import {GatheredBy} from '../list/gathered-by';
-import {GarlandToolsService} from '../../core/api/garland-tools.service';
-import {HtmlToolsService} from '../../core/html-tools.service';
 import {DeserializeAs} from '@kaiu/serializer';
 import {Craft} from './craft';
